requirements.txt
flask==3.0.2
google-generativeai==0.3.2
groq==0.4.2
python-dotenv==1.0.1
reportlab==4.1.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
matplotlib==3.8.2
redis==5.0.1
sqlalchemy==2.0.27
alembic==1.13.1
pdfkit==1.0.0
deep-translator==1.11.4
beautifulsoup4==4.12.3
requests==2.31.0
gunicorn==21.2.0 